tts_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
TTS==0.21.1
numpy==1.22.0
torch==2.1.2
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
python-multipart==0.0.6
soundfile==0.12.1 